Create and implement ReceivedFile
This commit creates and implements the ReceivedFile class. ReceivedFile acts as a file for a packet. So if we are getting 3 files from the server then there would be 3 received file objects. PacketManager decides which packet should go to which ReceivedFile object. If received file gets a header packet, it grabs the file name. If it gets a data packet, ReceivedFile puts in a sorted map, with the key being the packet number and the value being the data.

Co-authored-by: Dante Miller <mill8622@morris.umn.edu>